fix: failing cypress test, rename files
